feat: Program to experiment with perfect hashing
A potential method for identifying words in the list and detecting which
of them are subwords of a target word is to count the occurrences of
each letter a-z in the word then form the product of prime powers, where
each power is the number of occurrences of the corresponding letter.

This can be seen as a form of perfect hashing, since two words will have
the same hash iff they are anagrams of each other.

Calculating if one word is a subword of another is as simple as
computing whether the hash of the first divides the hash of the other.